Move ads upto gamescreen
